django_blog_it_en/requirements.txt
boto
django-simple-pagination
django-storages
Pillow
requests
gunicorn
whitenoise
dj-database-url
psycopg2
microurl
